Various changes
Clean the release artifacts prior to building. This ensures that the
build artifacts properly reflect the source.

Do not double-compress distsets.

Provide a revision.txt file that contains the git commit as given by
`git rev-list -n 1 HEAD`.

Do not attempt to publish to a mirror if no mirror is set.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>