Teach the build scripts how to skip building artifacts
Performing full builds during development of these scripts can chew up a
large chunk of time. Given that the process of building the artifacts
takes the majority of the time, but don't change much, provide a -S
option to skip building the artifacts. Uppercase is used as an
additional measure to prevent footshooting.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#8